+++ client-sdk/ts-web/core/docs/changelog.md
@@ -6,6 +6,8 @@ New features:
 
 - Added `signature.visitMessage` for use in code that looks at a message
   before signing.
+- Added `consensus.visitTransaction` for the case where that message to sign
+  is a consensus transaction.
 
 ## v0.1.0-alpha4